Area zones, director overhaul, player weapon slots
The ammo crate and health locker have also been merged with the main pack in the add-ins folder

Area zones is the new definition for radius teleporting, area zones are spots for sub bricks like items to spawn, and determine what zombie type should also spawn in the desired area.

This comes with a new tagging system, naming bricks with _Horde or _Special will change the spawn type of that brick to match those needed.

Item bricks with a tag from the itemscavenge.txt list will spawn what is the desired item.

There's also a new event called onAZfirstEntry for the main area zone brick to use, it's for calling specific events.

The director has seen some considerable changes, most importantly is that it no longer requires a hole bot brick to be placed in the map, previously, the bots would spawn to the desired location deemed from the player's placed brick, it is now done with the area bricks themselves.

There is a new factor that checks for players that are far ahead of the pack, if players ahead decide to rush, they will be greeted larger hordes and specials to prevent them from moving.

Player slots are a new system to prevent having more than one weapon or item type, the first slot are for weapons, second for secondaries, third for grenade, 4th for medical, 5th for secondary medical.

There are more features too, such as the director determining how the weapon item should spawn, but this is currently a bit bugged and does not work as I intended... Should look into this, there probably some other major changes I did but I cannot list them all.