2023 time
I need to publish something to get this whole project going again

First person bodies are a thing, and I'm using mounted images for the hats. So far it seems to work okay but I plan to integrate more hats in the future.

A whole bunch of misc changes too.